requirements.txt
# requirements.txt

fastapi==0.104.1
uvicorn==0.24.0
google-generativeai==0.3.0
requests==2.31.0
python-dotenv==1.0.0
elevenlabs==0.2.1
pydantic==2.5.0
